拆分成三個java檔
