+++ StaticFileHandler.java
@@ -1,0 +1,38 @@
+import com.sun.net.httpserver.HttpExchange;
+import com.sun.net.httpserver.HttpHandler;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
+public class StaticFileHandler implements HttpHandler {
+    private final String staticDir;
+
+    public StaticFileHandler(String staticDir) {
+        this.staticDir = staticDir;
+    }
+
+    @Override
+    public void handle(HttpExchange exchange) throws IOException {
+        if ("GET".equals(exchange.getRequestMethod())) {
+            String path = exchange.getRequestURI().getPath();
+            if (path.equals("/")) {
+                path = "/index.html";
+            }
+
+            Path filePath = Paths.get(staticDir + path);
+            if (Files.exists(filePath)) {
+                byte[] content = Files.readAllBytes(filePath);
+                exchange.sendResponseHeaders(200, content.length);
+                try (OutputStream os = exchange.getResponseBody()) {
+                    os.write(content);
+                }
+            } else {
+                exchange.sendResponseHeaders(404, -1);
+            }
+        } else {
+            exchange.sendResponseHeaders(405, -1);
+        }
+    }
+}
